move not used functions into an archive
