+++ pkg/archive/common/configmap.go
@@ -1,0 +1,57 @@
+/*
+Copyright 2020 Red Hat
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package common
+
+import (
+	"context"
+	"fmt"
+
+	"k8s.io/apimachinery/pkg/types"
+
+	common "github.com/openstack-k8s-operators/lib-common/pkg/common"
+
+	corev1 "k8s.io/api/core/v1"
+	k8s_errors "k8s.io/apimachinery/pkg/api/errors"
+)
+
+// CreateOrGetCustomConfigMap -
+func CreateOrGetCustomConfigMap(
+	ctx context.Context,
+	r ReconcilerCommon,
+	configMap *corev1.ConfigMap,
+) (string, error) {
+	// Check if this configMap already exists
+	foundConfigMap := &corev1.ConfigMap{}
+	err := r.GetClient().Get(ctx, types.NamespacedName{Name: configMap.Name, Namespace: configMap.Namespace}, foundConfigMap)
+	if err != nil && k8s_errors.IsNotFound(err) {
+		r.GetLogger().Info("Creating a new ConfigMap", "ConfigMap.Namespace", configMap.Namespace, "ConfigMap.Name", configMap.Name)
+		err = r.GetClient().Create(ctx, configMap)
+		if err != nil {
+			return "", err
+		}
+	} else {
+		// use data from already existing custom configmap
+		configMap.Data = foundConfigMap.Data
+	}
+
+	configMapHash, err := common.ObjectHash(configMap)
+	if err != nil {
+		return "", fmt.Errorf("error calculating configuration hash: %v", err)
+	}
+
+	return configMapHash, nil
+}
