+++ pkg/archive/common/env.go
@@ -1,0 +1,40 @@
+/*
+Copyright 2020 Red Hat
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package common
+
+import (
+	common "github.com/openstack-k8s-operators/lib-common/pkg/common"
+	corev1 "k8s.io/api/core/v1"
+)
+
+// +kubebuilder:object:generate:=false
+
+// EnvDownwardAPI - set env from FieldRef->FieldPath, e.g. status.podIP
+func EnvDownwardAPI(field string) common.EnvSetter {
+	return func(env *corev1.EnvVar) {
+		if env.ValueFrom == nil {
+			env.ValueFrom = &corev1.EnvVarSource{}
+		}
+		env.Value = ""
+
+		if env.ValueFrom.FieldRef == nil {
+			env.ValueFrom.FieldRef = &corev1.ObjectFieldSelector{}
+		}
+
+		env.ValueFrom.FieldRef.FieldPath = field
+	}
+}
