+++ pkg/archive/common/secret.go
@@ -1,0 +1,61 @@
+/*
+Copyright 2020 Red Hat
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package common
+
+import (
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+const (
+	// BITSIZE -
+	BITSIZE   int    = 4096
+	sshConfig string = `Host *
+    User cloud-admin
+    StrictHostKeyChecking no`
+)
+
+// SSHKeySecret - func
+func SSHKeySecret(name string, namespace string, labels map[string]string) (*corev1.Secret, error) {
+
+	privateKey, err := GeneratePrivateKey(BITSIZE)
+	if err != nil {
+		return nil, err
+	}
+
+	publicKey, err := GeneratePublicKey(&privateKey.PublicKey)
+	if err != nil {
+		return nil, err
+	}
+
+	privateKeyPem := EncodePrivateKeyToPEM(privateKey)
+
+	secret := &corev1.Secret{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      name,
+			Namespace: namespace,
+			Labels:    labels,
+		},
+		Type: "Opaque",
+		StringData: map[string]string{
+			"identity":        privateKeyPem,
+			"authorized_keys": publicKey,
+			"config":          sshConfig,
+		},
+	}
+	return secret, nil
+}
